XJ/requirements.txt
Flask
gunicorn
pyodbc
Pillow
